Rename to `VectorQueryTable`
