+++ src/components/VectorQueryTable/index.tsx
@@ -1,0 +1,6 @@
+import VectorQueryTable, {
+  VectorSimiliartyTableProps,
+} from "./VectorQueryTable";
+
+export default VectorQueryTable;
+export type { VectorSimiliartyTableProps };
