--- src/components/VectorSimilarityTable/index.tsx
@@ -1,6 +1,0 @@
-import VectorSimilarityTable, {
-  VectorSimiliartyTableProps,
-} from "./VectorSimilarityTable";
-
-export default VectorSimilarityTable;
-export type { VectorSimiliartyTableProps };
